requirements.txt
anthropic
requests
python-dotenv
typer